+++ sdk/storagecache/azure-resourcemanager-storagecache/src/main/java/com/azure/resourcemanager/storagecache/StorageCacheManager.java
@@ -27,6 +27,7 @@ import com.azure.resourcemanager.storagecache.fluent.StorageCacheManagementClien
 import com.azure.resourcemanager.storagecache.implementation.AmlFilesystemsImpl;
 import com.azure.resourcemanager.storagecache.implementation.AscOperationsImpl;
 import com.azure.resourcemanager.storagecache.implementation.AscUsagesImpl;
+import com.azure.resourcemanager.storagecache.implementation.AutoExportJobsImpl;
 import com.azure.resourcemanager.storagecache.implementation.CachesImpl;
 import com.azure.resourcemanager.storagecache.implementation.ImportJobsImpl;
 import com.azure.resourcemanager.storagecache.implementation.OperationsImpl;
@@ -39,6 +40,7 @@ import com.azure.resourcemanager.storagecache.implementation.UsageModelsImpl;
 import com.azure.resourcemanager.storagecache.models.AmlFilesystems;
 import com.azure.resourcemanager.storagecache.models.AscOperations;
 import com.azure.resourcemanager.storagecache.models.AscUsages;
+import com.azure.resourcemanager.storagecache.models.AutoExportJobs;
 import com.azure.resourcemanager.storagecache.models.Caches;
 import com.azure.resourcemanager.storagecache.models.ImportJobs;
 import com.azure.resourcemanager.storagecache.models.Operations;
@@ -61,6 +63,8 @@ import java.util.stream.Collectors;
  */
 public final class StorageCacheManager {
     private AmlFilesystems amlFilesystems;
+
+    private AutoExportJobs autoExportJobs;
 
     private ImportJobs importJobs;
 
@@ -246,7 +250,7 @@ public final class StorageCacheManager {
                 .append("-")
                 .append("com.azure.resourcemanager.storagecache")
                 .append("/")
-                .append("1.0.0");
+                .append("1.0.0-beta.1");
             if (!Configuration.getGlobalConfiguration().get("AZURE_TELEMETRY_DISABLED", false)) {
                 userAgentBuilder.append(" (")
                     .append(Configuration.getGlobalConfiguration().get("java.version"))
@@ -302,6 +306,18 @@ public final class StorageCacheManager {
             this.amlFilesystems = new AmlFilesystemsImpl(clientObject.getAmlFilesystems(), this);
         }
         return amlFilesystems;
+    }
+
+    /**
+     * Gets the resource collection API of AutoExportJobs. It manages AutoExportJob.
+     * 
+     * @return Resource collection API of AutoExportJobs.
+     */
+    public AutoExportJobs autoExportJobs() {
+        if (this.autoExportJobs == null) {
+            this.autoExportJobs = new AutoExportJobsImpl(clientObject.getAutoExportJobs(), this);
+        }
+        return autoExportJobs;
     }
 
     /**